Couse home link issue using $wwwroot
When using HTTPS for login only the $wwwroot is pointing to the https version of the site in the menunav.php, this is different to how all the other menu links work. Changing the course home links to use $CFG->wwwroot (as the other links use) appears to fix the issue.